+++ Main/testes.py
@@ -1,0 +1,3 @@
+among = ['Among Us',"https://download1478.mediafire.com/r8iwm6uij49gD-5rw1xygwaEQMDIx082wJqvwR22VajYQTfNr-JDegmR_kzR0O-FPgWQUWT5u4_jOUoOtrGLVWmNFEvJQGWsaLpWiHdv3vG3IjbuyZS6qfJ2F4dDeR1GJdhLXbxcEX-bYUNw3ka7Z_8woS3gJMH548Wh3dQqHWtplN8/t9i81p300f7ay5c/Among+Us.zip", r"downloads\Among Us"]
+
+print(among[0].replace(" ","") + '.zip')
